pyproject.toml
[project]
name = "advance-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.9",
    "python-dotenv>=1.1.0",
]
